This commit adds two things:
1. All options can now be specified in the webpack.config.js file.
This means that things like python version, folder, sourcemaps, etc.
can be overridden when needed.  In particular, being able to modify
the switches is useful.

2. A new option `keep_compiled` keeps the compiled files instead
of unlinking (deleting) them.  The default of this new switch
is the behavior of the module previously.

None of these changes require anything to be done to the config
files in projects already using the module.  Fully backwards-compatible.